+++ samples/WesternMenuPlayerExtension.cpp
@@ -8,7 +8,7 @@
 #include "limonAPI/LimonConverter.h"
 
 void WesternMenuPlayerExtension::processInput(const InputStates &inputState, const PlayerExtensionInterface::PlayerInformation &playerInformation [[gnu::unused]],
-                                              long time) {
+                                              uint32_t time) {
 
     static constexpr float PI = 3.14159265358979f;
